Move in_core and steady_state_only functions to helper.py and create new core_steady_state_mean function
